requirements.txt
flask>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
sqlalchemy>=1.4.0
